+++ app/src/test/java/com/termux/terminal/ResizeTest.java
@@ -93,7 +93,7 @@ public class ResizeTest extends TerminalTestCase {
 		enterString("\033[2J");
 		for (int r = 0; r < rows; r++) {
 			for (int c = 0; c < cols; c++) {
-				int style = getStyleAt(r, c);
+                long style = getStyleAt(r, c);
 				assertEquals(119, TextStyle.decodeForeColor(style));
 				assertEquals(129, TextStyle.decodeBackColor(style));
 			}
@@ -105,7 +105,7 @@ public class ResizeTest extends TerminalTestCase {
 		// After resize, screen should still be same color:
 		for (int r = 0; r < rows - 2; r++) {
 			for (int c = 0; c < cols; c++) {
-				int style = getStyleAt(r, c);
+                long style = getStyleAt(r, c);
 				assertEquals(119, TextStyle.decodeForeColor(style));
 				assertEquals(129, TextStyle.decodeBackColor(style));
 			}
@@ -116,7 +116,7 @@ public class ResizeTest extends TerminalTestCase {
 		resize(cols, rows);
 		for (int r = 0; r < rows; r++) {
 			for (int c = 0; c < cols; c++) {
-				int style = getStyleAt(r, c);
+                long style = getStyleAt(r, c);
 				assertEquals(119, TextStyle.decodeForeColor(style));
 				assertEquals("wrong at row=" + r, r >= 3 ? 200 : 129, TextStyle.decodeBackColor(style));
 			}
